Move `Field` location and coordinates members to `FieldData`
