+++ src/field/field.cxx
@@ -32,89 +32,8 @@
 #include <utils.hxx>
 #include <bout/mesh.hxx>
 
-namespace bout {
-/// Make sure \p location is a sensible value for \p mesh
-///
-/// Throws if checks are enabled and trying to use a staggered
-/// location on a non-staggered mesh
-CELL_LOC normaliseLocation(CELL_LOC location, Mesh* mesh) {
-  AUTO_TRACE();
-
-  // CELL_DEFAULT always means CELL_CENTRE
-  if (location == CELL_DEFAULT) {
-    return CELL_CENTRE;
-  }
-
-  // No mesh means we can't check if we're using staggered grids, so
-  // we'll have to trust the user in this case. This can happen if
-  // we're making a field before the global mesh has been initialised
-  // -- probably not good, but possible.
-  if (mesh == nullptr) {
-    return location;
-  }
-
-  if (mesh->StaggerGrids) {
-    if (location == CELL_VSHIFT) {
-      throw BoutException(
-          "Field: CELL_VSHIFT cell location only makes sense for vectors");
-    }
-    return location;
-  } else {
-#if CHECK > 0
-    if (location != CELL_CENTRE) {
-      throw BoutException("Field: Trying to set off-centre location on "
-                          "non-staggered grid\n"
-                          "         Did you mean to enable staggered grids?");
-    }
-#endif
-    return CELL_CENTRE;
-  }
-}
-} // namespace bout
-
 Field::Field(Mesh* localmesh, CELL_LOC location_in, DirectionTypes directions_in)
-    : FieldData(localmesh == nullptr ? bout::globals::mesh : localmesh),
-      location(bout::normaliseLocation(location_in, fieldmesh)),
-      directions(directions_in) {
-
-  // Need to check for nullptr again, because the fieldmesh might still be
-  // nullptr if the global mesh hasn't been initialized yet
-  if (fieldmesh != nullptr) {
-    // sets fieldCoordinates by getting Coordinates for our location from
-    // fieldmesh
-    getCoordinates();
-  }
-}
-
-void Field::setLocation(CELL_LOC new_location) {
-  AUTO_TRACE();
-
-  location = bout::normaliseLocation(new_location, getMesh());
-
-  fieldCoordinates.reset();
-  // Sets correct fieldCoordinates pointer and ensures Coordinates object is
-  // initialized for this Field's location
-  getCoordinates();
-}
-
-CELL_LOC Field::getLocation() const {
-  AUTO_TRACE();
-  return location;
-}
-
-Coordinates* Field::getCoordinates() const {
-  auto fieldCoordinates_shared = fieldCoordinates.lock();
-  if (fieldCoordinates_shared) {
-    return fieldCoordinates_shared.get();
-  }
-  fieldCoordinates = getMesh()->getCoordinatesSmart(getLocation());
-  return fieldCoordinates.lock().get();
-}
-
-Coordinates *Field::getCoordinates(CELL_LOC loc) const {
-  if (loc == CELL_DEFAULT) return getCoordinates();  
-  return getMesh()->getCoordinates(loc);
-}
+    : FieldData(localmesh, location_in), directions(directions_in) {}
 
 int Field::getNx() const{
   return getMesh()->LocalNx;
