+++ src/field/field_data.cxx
@@ -7,9 +7,64 @@
 #include <field_factory.hxx>
 #include "unused.hxx"
 
+namespace bout {
+/// Make sure \p location is a sensible value for \p mesh
+///
+/// Throws if checks are enabled and trying to use a staggered
+/// location on a non-staggered mesh
+CELL_LOC normaliseLocation(CELL_LOC location, Mesh* mesh) {
+  AUTO_TRACE();
+
+  // CELL_DEFAULT always means CELL_CENTRE
+  if (location == CELL_DEFAULT) {
+    return CELL_CENTRE;
+  }
+
+  // No mesh means we can't check if we're using staggered grids, so
+  // we'll have to trust the user in this case. This can happen if
+  // we're making a field before the global mesh has been initialised
+  // -- probably not good, but possible.
+  if (mesh == nullptr) {
+    return location;
+  }
+
+  if (mesh->StaggerGrids) {
+    if (location == CELL_VSHIFT) {
+      throw BoutException(
+          "Field: CELL_VSHIFT cell location only makes sense for vectors");
+    }
+    return location;
+  } else {
+#if CHECK > 0
+    if (location != CELL_CENTRE) {
+      throw BoutException("Field: Trying to set off-centre location on "
+                          "non-staggered grid\n"
+                          "         Did you mean to enable staggered grids?");
+    }
+#endif
+    return CELL_CENTRE;
+  }
+}
+} // namespace bout
+
+FieldData::FieldData(Mesh* localmesh, CELL_LOC location_in)
+    : fieldmesh(localmesh == nullptr ? bout::globals::mesh : localmesh),
+      location(bout::normaliseLocation(
+          location_in, fieldmesh)) { // Need to check for nullptr again, because the
+                                     // fieldmesh might still be
+  // nullptr if the global mesh hasn't been initialized yet
+  if (fieldmesh != nullptr) {
+    // sets fieldCoordinates by getting Coordinates for our location from
+    // fieldmesh
+    getCoordinates();
+  }
+}
+
 FieldData::FieldData(const FieldData& other) {
   copyBoundary(other);
   fieldmesh = other.fieldmesh;
+  location = other.location;
+  fieldCoordinates = other.fieldCoordinates;
 }
 
 FieldData::~FieldData() {
@@ -23,6 +78,8 @@ FieldData::~FieldData() {
 FieldData& FieldData::operator=(const FieldData& other) {
   copyBoundary(other);
   fieldmesh = other.fieldmesh;
+  location = other.location;
+  fieldCoordinates = other.fieldCoordinates;
   return *this;
 }
 
@@ -95,4 +152,38 @@ Mesh* FieldData::getMesh() const {
     // done.
     return bout::globals::mesh;
   }
+}
+
+FieldData& FieldData::setLocation(CELL_LOC new_location) {
+  AUTO_TRACE();
+
+  location = bout::normaliseLocation(new_location, getMesh());
+
+  fieldCoordinates.reset();
+  // Sets correct fieldCoordinates pointer and ensures Coordinates object is
+  // initialized for this Field's location
+  getCoordinates();
+
+  return *this;
+}
+
+CELL_LOC FieldData::getLocation() const {
+  AUTO_TRACE();
+  return location;
+}
+
+Coordinates* FieldData::getCoordinates() const {
+  auto fieldCoordinates_shared = fieldCoordinates.lock();
+  if (fieldCoordinates_shared) {
+    return fieldCoordinates_shared.get();
+  }
+  fieldCoordinates = getMesh()->getCoordinatesSmart(getLocation());
+  return fieldCoordinates.lock().get();
+}
+
+Coordinates* FieldData::getCoordinates(CELL_LOC loc) const {
+  if (loc == CELL_DEFAULT) {
+    return getCoordinates();
+  }
+  return getMesh()->getCoordinates(loc);
 }
